web/requirements.txt
flask
firebase-admin
python-dotenv